+++ src/lib.rs
@@ -84,7 +84,7 @@ use core::fmt;
 /// [`WindowHandle::Xlib`] on macOS, it would just be weird, and probably
 /// requires something like XQuartz be used).
 #[non_exhaustive]
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
 pub enum WindowHandle<'window> {
     /// A raw window handle for UIKit (Apple's non-macOS windowing library).
     ///
@@ -206,7 +206,7 @@ pub enum WindowHandle<'window> {
 /// [`RawDisplayHandle::Xlib`] on macOS, it would just be weird, and probably
 /// requires something like XQuartz be used).
 #[non_exhaustive]
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
 pub enum DisplayHandle<'display> {
     /// A raw display handle for UIKit (Apple's non-macOS windowing library).
     ///